Improvement: Output files are stored in their own directory
Output files are stored in their own directory to structurize the
results of the experiments. The directory name can be changed from the
commandline. I have also created a large part of the logic to extract
git history from files and use it as part of the context to the LLM.