Feature: Add pylint and astroid to statically analyze files
Added functionality to enable static analyzation of code to provide more
context to the llm. Made a few changes to the git stats functions. The functions
no longer normalize and create scoring values, but rather sends all info about
churn, number of commits etc. to the llm. Instead we provide some context about
the what is a high/low value for a given variable and what that means.