Feature: Design and implementation of the baseline
The majority of the base line is now complete. We have moved on from using
weighted scores to give git history and other stats. Instead they are processed in a
NL-format (according to research, this is better for the LLM to understand).

I have implemented a ranking script that calculcates the ndcg, kendall-tau, rbo
and spearman score to measure the similarity of the output to the ground truth given
that the output is in the correct format. I have implemented formatting functions
that can fix small formatting changes, but these fixes do not cover hallucinations
from the llm.

Made changes to the README, and added a requirements.txt file to make setup simpler.