Refactor: Redesigned the project architecture
The project architecture was refactored to a python package to make it more
scalable for future features and for others to work on the project in the future.
Added a pyproject file and moved my code into a src folder so that it is
seperated from the python_smell_detector and experiments directories in the root dir.